Add comprehensive tests for Octree and Tetree spatial indexing
Introduced detailed JUnit tests for Octree and Tetree, covering insertion, retrieval, geometric consistency, and space-filling curve properties. These tests ensure accuracy across edge cases, coordinate quantization levels, partitioning, and hierarchical spatial relationships.

+++ lucien/src/main/java/com/hellblazer/luciferase/lucien/Tetree.java
@@ -71,7 +71,7 @@ public class Tetree<Content> {
      * @param linearIndex - the index in the space filling curve
      * @return the Content at the linear index
      */
-    public Content get(int linearIndex) {
+    public Content get(long linearIndex) {
         return contents.get(linearIndex);
     }
